Implement base TreeBuilder
A TreeBuilder allows various types of entities to take more than
a single Chunk in the pool.  The builder creates as many levels
of indirection as are necessary to represent the entire tree of
data as a single chunk.	 It is parameterizable based on the names
used for the base chunks and the indirection chunks.

This change implements the TreeBuilder API and enough code to
build a "tree" containing a single element.

aegis-delta: jpool.1.0.D019
aegis-uuid: 2273243b-81ef-494e-8e28-966e80224d6c