Delete all local refs on file read.
The file read loop misses deleting a local ref, causing a space
leak on large files.

This is a difficult bug to write a test for, since it merely
causes the entire contents of a file to be read into memory while
the file is being read.	 To avoid needing to write huge files,
this test is exempt against the baseline, but has been tested
during development against large files to demonstrate the fix.

aegis-delta: jpool.1.0.D035
aegis-uuid: 01628e40-d4dd-48be-bb89-05ec1c4d00d0