Logger messages print Array(...) for %s in message.
See the DevMap test as non-root for an example.

This is looking like a compiler bug.  It seems to do the wrong
thing when the generated code goes through a proxy for the
logging functions rather than calling them directly.

A workaround seems to be to no make the logging calls protected.

aegis-delta: jpool.1.0.D057
aegis-uuid: 5f650ee8-27b7-4e00-bb47-0215e57aff4b