Create dirstore.
Implement a dirstore that maps more closely to how Posix stores
directories.  Right now, we're storing the filename in the inode,
which is messing up hardlinks.	Create a new class, similar to
the TreeBuilder (using it, in fact) that manages the simplified
directory we need.  It stores a sorted pairing of filenames and
the hash of their node information in one or more chunks
referencable by the Tree walker.

aegis-delta: jpool.1.0.D068
aegis-uuid: 8248ec8f-317a-404b-b022-80fd306dd697