Move node names out of the 'Attributes' entries and into the directory itself.
Having the filename in the Attributes is causing semantic
mismatch with the Posix filesystem where the attributes are
stored in the inode and are fairly independent of the name they
are linked to.	This shows up primarily with multiply-hardlinked
files where the same inode is referred to by different names.
This fixes some serious problems archiving hardlinked files.

aegis-delta: jpool.1.0.D069
aegis-uuid: 5138083e-cb06-4c6e-a29d-2a6534e37ce2