Convert SeenDb to per-pnode table.
Instead of storing a database record for every file in every
directory, only store a record for each directory, and use a BLOB
column to hold a packed representation of the entries in that
directory.  This gets rid of a lot of wasted space in the
database and makes the index significantly smaller.

aegis-delta: jpool.1.0.D080
aegis-uuid: 0ad7f155-8e89-44f5-a282-90499fb14847