Allow reuse of native code from SBT.
SBT invokes the test suite using a new class loader each time.  The
JVM doesn't allow what it thinks is the same shared library to be
loaded from different classloaders.

To work around this, copy the shared library into a temp file, and use
this temp file as the library to load.  This might cause a small
memory leak, but allows the shared lib to be reinitialized and new
versions tested without having to exit the JVM.