Convert Progress meter from object to class.
Eliminate all uses of the Progress singleton object, and properly
open/close the meter passing it around as needed.  The real meter is
still a singleton calls TheProgress, which is not referenced outside
of open().  As such, it is still erroneous to have more than one
progress meter open, but all of the client code should now be able to
handle this.