Add 'forced' hint to progress meter state.
The 'formatState' method in a given progress meter can be called both
periodically, as well as at certain "forced" times (such as when log
messages are printed, or when the meter is unregistered).  Add an
argument to the method to indicate whether the update has been forced
or not.

This will be useful when performing rate computations, and such.