Simple implementation of 'Clone'.
The Clone operation walks through the entire tree representing a
snapshot, copying it to a destination pool.

This implementation should not be used, since it uses a pre-order
traversal to write the data.  If there is an interruption in the
backup, the destination pool will contain entries that refer to nodes
that haven't been copied yet.  Even worse is that subsequent copies
will not copy these nodes, since the clone assumes that the presence
of a node indicates all of its children are present.