Fix traversal on GC.
Eliminates the problem with aborted clones by fixing the traversal
order in the garbage collector.  The visitor object passed around now
has two methods, one called pre-order in the search to check if a node
has been "marked", and a post-order call to actually mark the node.

The terms might be a little misleading from a traditional GC, since we
aren't concerned with cycles.  Cycles could easily be supported by
doing the actual mark in the isSeen() call, and having 'mark' do a
copy.  But, backup data should never contain cycles, or restore won't
ever terminate.