Simplify nested package declarations.
In scala 2.7.x, saying 'package a.b.c' would perform an equivalent of
'import a._', and 'import a.b._', as well as making everything in the
'c' package visible.  This is no longer the case with 2.8, and only
the names from package 'c' are visible.

Previously, the 2.8 port added a bunch of extra import statements to
bring in these packages for the parent packages.

The proper idiomatic way of doing this now in Scala is to say
something like:

  package org.davidb.jpool
  package tools

The first will declare the package org.davidb.jpool, and make
everything in org.davidb.jpool._ visible.  The second will then set
the package to org.davidb.jpool.tools, bringing those names as well,
without removing the names brought in by the first package
declaration.

Replace all of the extra imports of parent packages with multi-line
package declarations.