Fix expiration of SeenDb entries.
Previously, the seen database tried to store the minimum expiration
time, but it was being computed incorrectly, often resulting in this
field being -1.  In addition, minimum expiration isn't really useful
as an expiration value.

Change this to compute the maximum expiration time (correctly) of all
entries in a given directory.  This makes it a simple query that if
the maximum expiration is older than the current time, then the entire
directory can be purged.

Directories that are still active will have their old entries purged
every time they are updated.  Directories that are inactive (such as
deleted directories) will be purged entirely when their latest
expiration expires.  This will still leave some expired entries, but
will keep things fairly clean.

Also adds an index on the expiration time.