Crypto: Implement encrypted Chunk Read/Write
The EncryptedPoolFile generates secret keys as needed, encryptes them
with the backup public key, and uses that to encrypt the backup data.
Upon read, it will use the secret to decrypt the secret to be able to
read the chunks.

The secrets store the key fingerprint of the backup secret to make
sure that the correct key is used to attempt restore.