Implement encrypted backups
The KeyGen tool can generate a new key for a pool.  The pool must have
no data in it.  It is also possible to copy a certificate from another
pool to an empty pool to support write-only encrypted backups.  Note
that if a backup fails, there is no way to do index recovery without
the private key.