Switch to plaintext metadata
Since commit 761f387af1a60404d7a285fab0818f4db498527a, "Write backup
metadata to text files" jpool has stored all of the backup metadata in
text files as well as the SQL database.

The SQL database is unnecessary for what amounts to a set of
properties and a list of hashes.

Remove the SQL backend for the metadata, and read and write the fields
from the text files.

If the user encounters an old backup that doesn't have the text files,
choices are to either run a jpool one commit before this one to create
them, or to remove the metadata directory entirely and recreate it by
scanning the backup pool files directly.  Jpool doesn't yet make use
of the UUID in the properties, so it doesn't hurt anything to change
it.