Faster reading of only chunk headers
Certain recovery operations need to skim the headers of the chunks.
Speed this up, at least for un-encrypted pool files by just reading in
the header instead of decoding the entire chunk.  This adds a new
method to pool files.

Encrypted reads could also be sped up, but currently decode the entire
chunk and return the header information.