Per file indexing
Implement an indexer that only indexes the nodes that are found in a
single pool file.  The format is inspired by the pack index format
used by git.  There is an array of offsets based on the first byte of
the Hash, which is used to bound the binary search.  This tends to
make most lookups quite quick.  Since these indices don't have to
store which file a given value is in, they take less space.  This is
countered by needing to search through multiple index files to find a
node.

The advantage will be that since each pool file has its own index,
operations such as garbage collection can replace files without having
to recompute the entire index.

This change only adds the code to read and write these index files as
well as a unit test to verify it works.