JIT decode of backup index
Instead of decoding the backup index on startup, leave the encoded
data in memory, and decode the segments as needed.  There are two
advantages to this: startup time is significantly faster, since most
data is not decoded; and the memory footprint is much smaller, since
the decoded buffers have a lot of overhead.

Since the encoded buffers are mmapped, they are also only paged in as
needed, which also reduces startup time and memory footprint for many
cases.