Reenable mmap prefilling of index
Although prefilling slows down loading when most objects aren't going
to be visited, the thrashing from out-of-order loading when the pages
are accessed out of order costs significantly more.  Since most uses
are the pool are going to access scattered objects, it seems more
useful to preload and have a longer startup.