Port project to SBT 0.10
The native compilation happens in a similar manner, and seems to work
reasonably well.  There are several hardcoded paths in it, however.

SBT 0.10 appears to run the scaltests tests in parallel.  Since many
of the tests involve state and/or files, put a lock around any test
that uses a pool.  This restricts most of the tests to be single
threaded.

The jpool wrapper script isn't ported to the new directory layout.

If sbt is killed, the native files, and various temp directories
aren't cleaned up properly.