Migration to sbt 0.11.0-RC1
This version of sbt keeps the scala compiler and library in a common
directory, instead of copying it to each project.  Although more
efficient, it means we need to be able to run based on the proper
classpath.

Add a command to sbt 'mkrun' that will write a small script
'target/classpath.sh' containing the class path necessary to run
programs.