Native workaround for JDK7 bug
A bug in some versions of the JDK7 prevent jpool from building:
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=7185778
http://stackoverflow.com/questions/10669704/error-with-javah-when-writing-jni-in-scala-with-java-7

The 'javah' tool doesn't accept class names that contain '$'
characters.  To work around this, put all of the native methods in a
regular class, and inherit that class into it's companion singleton
object.  This will remove the '$' from all of the external names.